Copy the latest version of the project

+++ app/models/Reading.java
@@ -10,9 +10,9 @@ public class Reading extends Model
   public int code;
   public double temperature;
   public double windSpeed;
-  public int pressure;
+  public double pressure;
 
-  public Reading(int code, double temperature, double windSpeed, int pressure)
+  public Reading(int code, double temperature, double windSpeed, double pressure)
   {
     this.code = code;
     this.temperature = temperature;
